Reorganized for standard vs internal tests. Added one standard test
